rachael updates

+++ manjaro/03-link-rc.sh
@@ -3,16 +3,21 @@
 REPO="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." >/dev/null 2>&1 && pwd )"
 
 cat > ~/.bashrc <<EOF
+[[ $- != *i* ]] && return
+
 source $REPO/bashrc
 source $REPO/bash_prompt
+source ~/.bashrc.local
 EOF
 
-for RC in inputrc tmux.conf xinitrc ; do
+touch ~/.bashrc.local
+
+for RC in inputrc tmux.conf ; do
     rm -f ~/.$RC
     ln -s $REPO/$RC ~/.$RC
 done
 
-for CONFIG in alacritty awesome gtk-3.0 lxterminal nvim wallpaper ; do
+for CONFIG in alacritty ; do
     rm -rf ~/.config/$CONFIG
     ln -s $REPO/config/$CONFIG ~/.config/$CONFIG
 done
@@ -29,4 +34,3 @@ cat > ~/.gitconfig <<EOF
 EOF
 
 mkdir -p ~/usr/bin
-ln -sf ${REPO}/usr/bin/pm-make ~/usr/bin/pm-make
